Extracting query parts to properly build result (#137)
Currently `ProcessSelectQuery` takes a string and that's a problem
because we have to extract
needed data from it like tableName, where clause, column names, etc ...
which means that we have to
reparse sql statement. 
This PR is just a kind of workaround to get these elements.
Proper and long-term solution would be have fully typed safe `query`
argument that would have already parsed elements.