Make async_search work (#148)
Screens of all 4 types of `async_search` are here in the second (last)
paragraph:

https://www.notion.so/Translation-Kibana-Clickhouse-f354e9139eed4d258c045d8a0ef198c8?pvs=13

Most of the code is still the 'I want it working ASAP', ugly style. I
wanted to merge it ASAP though to have a working version somewhere on
`main`. I'd improve the code in a separate PR tomorrow.