New compose CI file (#256)
We are creating separate docker-compose files for CI builds. It speeds
up 7->4minutes with potential for further improvements.
1. Use less containers (no `mitmproxy` or `clean`).
2. Improve smoke tests.

Unfortunately, docker-compose tends to be fragile, so please let's do a
minimal amount of changes in `yml`. In the past, separate folders and
other tweaks caused instabilities. Happy to experiment with the
following PRs.