Straighten '"' handling and index name patterns when parsing table names (#298)
We've had a couple of workarounds to handle the presence of `"` in the table name. Unfortunately, once we introduced configuration processing and name pattern matching, it was no longer serving its purpose.

I unified the approach in the whole codebase - now table names are always passed without `"`. Those are added at the end when constructing `FROM` clause.