Add a failing test caused by JSON unmarshaller (#269)
Default JSON unmarshaller in Go parses every number to `float64`. It
causes rounding errors when number is a big integer. E.g. we send an
incorrect number to Clickhouse, not the one we received, which is proved
by this PR.

I'm commenting out this test for now, and I've added an issue to fix
that.