Quickfix to terms-enum (#429)
Fixes a panic in our new autocompletion feature. I fixed it there, but
I'd guess that the problem exists in all our selects and needs some
further research/general fix. Generally we don't handle null values
correctly, we assume that the value exists. Then we cast to that type
and get a panic.