Add a few search tests with `DateTime`s (#425)
I only don't test creating response. It's pretty trivial before the
`Aggregations` change and seems to work for a long time. There are lots
of tests for that in `Aggregations` branch, which will probably be
merged soon, and current code will be thrown away, so I see little value
in testing that.