Limit facets to 20000 rows (#507)
Before `facets` queried Clickhouse for all rows, now it's `20000`. I
tried `50000`, but it takes `0,2-0,3sec`, it doesn't feel so smooth.
With `20000` results are almost instant.
I think we'll be able to improve it so `50000` should be instant as
well. You can see that I query `SELECT * LIMIT 20000` now. In columnar
DB querying for `2` fields instead of `20` should be much faster, and we
can drastically limit number of fields that we query, but for that we
need to keep a list of fields in `WHERE` clause that we don't do at the
moment.
![Screenshot 2024-03-11 at 11 16
35](https://github.com/QuesmaOrg/poc-elk-mitmproxy/assets/5407146/56748a9f-61f6-4238-bec2-8ee43e1dfb78)