Quickfix create table parser (#514)
There was a small bug in parser: `Tuple(DateTime(3))` returned parsing
error, it needed to be `Tuple(DateTime(3) )` with a space. Because of
that `schema_loader` didn't load `kibana_sample_data_ecommerce` table.
It happened only in `~1/15` cases, so was hard to spot.
Also fixed a place where we'd panic in case of parsing error. Now it's a
warn.
Before:
![Screenshot 2024-03-11 at 16 08
31](https://github.com/QuesmaOrg/poc-elk-mitmproxy/assets/5407146/c5749d6d-0d08-40db-8740-730a61d28dc3)
After:
![Screenshot 2024-03-11 at 16 19
17](https://github.com/QuesmaOrg/poc-elk-mitmproxy/assets/5407146/a9ce3fc5-970d-4c9f-9b85-b2646e9d80c7)