Keep if column is nullable or not (#450)
**Old, semi-true motivation:** We need this information while processing
results from Clickhouse. Depending on column being nullable or not, the
type of field returned from Clickhouse differs.
**New motivation:** since it's already coded, it doesn't hurt to add
that. It might become useful, too. Also in our UI in `Schema` view,
types are now in more appropriate format. See screen below. Before it
didn't show `Nullable(...)`

Added this to both `schema_loader` and `parserCreateTable`, so it's set
in both cases.
Added tests for that.

![Screenshot 2024-03-25 at 11 02
21](https://github.com/QuesmaOrg/poc-elk-mitmproxy/assets/5407146/d955bd2b-0c2b-420f-98ce-3a1b6449dec1)