Fixed flakiness in `SELECT *` tests (#627)
For `SELECT *` we translate `*` by iterating over a map with columns ->
the order is nondeterministic.
Before I hacked that in tests by checking for a `.* field-one .*` regex
instead of `field-one, field-two, ...`, but it might hide some errors
via doing that.
So I changed that for a regex checking all permutations of fields, e.g.
`(a,b,c)|(a,c,b)|(b,a,c)| [...]`. Now if a test passes, we're 100% sure
the result was correct.

It doesn't make tests run slow when nr of fields is just a couple like
we have now.