Fix not using aliasing in a couple of places (#608)
In parsing, in a couple of places, we didn't take aliasing into account
at all. I went through entire code of the parser, and I think fixed all
such issues.
It'd be probably good to make a couple of tests for that, but it would
take some time and I'd like to finish a couple of small fixes I have in
my backlog before that. (edit: for one of them I'm already preparing one
aliasing test)

Before: on the right you can see error and a very weird SQL. It's like
that because it doesn't alias `timestamp -> @timestamp` properly and
tries to find it in `attributes`
![Screenshot 2024-03-26 at 08 39
24](https://github.com/QuesmaOrg/poc-elk-mitmproxy/assets/5407146/149a7d62-2034-4e9e-841e-2aa0b72e482f)
After: SQL is correct
![Screenshot 2024-03-26 at 09 04
21](https://github.com/QuesmaOrg/poc-elk-mitmproxy/assets/5407146/6b257dfa-599a-4f92-9407-5366745966a4)