Parse all requested fields (#740)
Expand query parser to include all field names that were requested during the search operation.

So far, we used `FieldName` for that, but this doesn't work anymore since anyone can request more than a single field. Leaving it there for the sake of backward compatibility in the near future. I will try to get rid of it soon.