Add basic Lucene support (#732)
Here you can find description of the language:
* (more concise, recommended)
https://lucene.apache.org/core/2_9_4/queryparsersyntax.html
*
https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-query-string-query.html

I called it basic support, as we agreed with Jacek to not support 100%
of the language. But still, a vast majority is supported, I think, + I
couldn't make parser to crash/work unexpectedly for a few
unsupported/incorrect inputs.
For more details, you can look at the tests + `lucene_parser.go` for a
list of not supported features.