Add more debug to search part 1 (#797)
* Added warn/error debug everywhere where something might go wrong in
`queryparser.query_parser.go`
* Unified error messages to start with lower case. While research for
loggers, I found out that it's considered best practice, and mixed
lowercase and uppercase would be less readable.

Wasn't completely sure in all cases what should be a warn and what an
error. We'll probably merge it as is, but in the future please let me
know/feel free to improve that yourself.