Add (very) basic support for first pipeline aggregation (#829)
I add a very basic support for our first aggregation: `bucket_script`.
Here, it basically means `return count()`. It's a pretty common
aggregation in OpenSearch Visualize, so I thought about implementing it
fully - it's not that difficult, but let's do it later if we want to.

I also:
* Added one `fullTextField` to the table, explained why in the comment
below.
* Improved support for `meta` field in the request. The dashboard didn't
work without it, even though `meta` is simply copied from a request to
the response. I didn't copy it in some cases, probably still don't do,
but let's deal with it once we see a testcase where it doesn't work yet.

Before:

![Elastic_and_Google_Calendar_-_Week_of_April_29__2024](https://github.com/QuesmaOrg/poc-elk-mitmproxy/assets/5407146/32cd530c-ed74-4400-be6a-6523e8d96305)

After:

![Elastic_and_Google_Calendar_-_Week_of_April_29__2024](https://github.com/QuesmaOrg/poc-elk-mitmproxy/assets/5407146/1f380f6f-9ec6-4da8-b231-1c21dbd81197)