Second part of reporting unsupported query types (#44)
Basically did the same thing as with aggregations. Tested receiving
every `Query DSL` query type possible, to see if we catch it. We seem
to. E.g. at the screens `shape` is a `Query DSL` unsupported query,
while `ip_range` is an aggregation - you can see both.

In the code I needed to only start returning if parsing went fine or not
for bool expressions, as they can contain subqueries with all other
`Query DSL` queries.
![Screenshot 2024-05-07 at 17 07
50](https://github.com/QuesmaOrg/quesma/assets/5407146/f2552446-f7fe-4e9d-a851-2c911bad95c5)
![Screenshot 2024-05-07 at 17 08
02](https://github.com/QuesmaOrg/quesma/assets/5407146/5aca5ea2-477f-4f9b-ab46-7cf5e3cf800b)