Remove no-op content-length override (#111)
In Go, you cannot set headers after a call to WriteHeader. Once WriteHeader is called, the HTTP status code and any set headers are sent to the client, and the HTTP response is considered to be in a "header written" state.

After this point, any attempts to set or modify headers will have no effect.