Typed SortFields instead of concatenated string (#169)
Introducing an intermediate representation of a single sorting order property. Until now, we'd parse into and pass around a string of the form `"property_name" order,` which we'd need to be parsed again later for the sake of response creation.

To avoid that, I introduced a dedicated type that carried information about a field and sorted orders. Later, this type is converted into an SQL-friendly string form. `SortField` itself does not know how to convert itself to SQL, but `QueryTranslators` do.