Use JSON or NDJSON types instead of []byte (#175)
This is part of the global change: Every `map[string]interface{}`,
`[]byte` and `string` that holds parsed JSON will be replaced with type
`JSON`.

This PR contains changes in the ingest part only. `search.go` is crowded
now.

Move types `JSON` and etc to the more generic `types` will done in next
PR.