Small fixes `top_metrics` (#188)
3 small fixes:
* `GROUP BY field1, field2` instead of `GROUP BY field1field2` 😄 
* `MAX` -> `maxOrNul`, like in all other places, just to be sure
* instead of `fieldname: value`, we returned `"windowed_fieldname":
value`. Fixed by unquoting if there are quotes (should always be, but
again, just to be sure)

 
Also added a type check in `date_histogram`. Should be unnecessary if
code is correct, but we used to panic when we had bugs before my fixes.


Fixes e.g. `Last value` view
Seems to be working, checked with Clickhouse, see `5` below

![ClickHouse_Query__SELECT_order_date__customer_id_FROM_kibana_sample_data_ecommerce_WHERE_order_date____2024-05-10_00_01_02_000__AND_order_date____2024-05-09_00_01_02_000__ORDER_BY_order_date_desc](https://github.com/QuesmaOrg/quesma/assets/5407146/40456bf7-8274-4986-83ee-bca920fa0ac3)
![Screenshot 2024-05-22 at 12 44
56](https://github.com/QuesmaOrg/quesma/assets/5407146/878236e2-4ad9-41f9-9445-ee7be6233fa5)