Query column as expression 3 (#234)
Removes `SuffixClauses`, and introduces `OrderBy []SelectColumn`, and
`limit int` instead.
This array of `OrderBy` is needed for sure, even my already more or less
finished PR with sizes introduces it. With multiple subaggregations,
even all simple `terms`, we need to do `ORDER BY` on multiple fields, so
just 1 `suffixClause` as it was before, is not enough.
Also removes `SortFields`, as what was there can now be simply stored in
this `OrderBy` with an advantage that it's properly typed, unlike
`SortFields`

There is one hackish function `*Query.OrderByFieldNames()`, but as it
seems to work, I'd merge it as it is to enable some further work on
schema, and maybe fix it sometime later.
I hope rest is more or less fine.