Quickfix incorrect tests (#290)
It's a bit concerning that these tests passed most of the time, even
though we expected 1 query in mock, but actually sent 2.
Maybe it's introduced by this newly added
`mock.MatchExpectationsInOrder(false)`? Maybe sqlmock's bug? Maybe my
bug?
I'll try to inspect that for a bit, but it seems to solve the flakiness
+ results should **now** be correct