Quesma Schema representation (#297)
Introduced an intermediate Quesma schema representation, which will eventually become the source of truth about data and their types. It defines its own types of "search types" which map independently to Elasticsearch/Opensearch and Clickhouse/Hydrolix.

This change is transparent, it introduces the concept, which is not yet used anywhere. In the next changes, I will start relying on it instead of `IsFullTextMatch` flag.

The set of recognized types is incomplete, naming might be confusing (SchemaLoader/SchemaRegistry), there's also excessive copying and suboptimal error handling. To be evened out incrementally.

Tests are coming after SchemaLoader/SchemaManagement refactoring - don't want to engage in SQLMock fun if I am able to stub.