Plugin registry - transformer pipelines (#336)
This is a proposal PR. It addresses:
- what is a plugin ?
- how plugins are registered?
- how  Quesma core uses "plugins"?

There are 3 plugins registered. Two of them are enabled. 

Indexes starting with "kibana_*" uses the newer column naming
convention. Field names have "."
Other indexes are uses the current method. 


Not sure if it should be merged as is. We should discuss Quesma core vs
plugins responsibility.