[Slight, not full fix] Arrays with different lengths for different aggregations (#345)
Before this dashboard below (from `sample_flights`) didn't work, in
`util/mergeMaps()` arrays were of different sizes, and we returned some
completely empty results because of that.
Now it more or less works:
![Screenshot 2024-06-17 at 16 30
24](https://github.com/QuesmaOrg/quesma/assets/5407146/c61d1d24-b0eb-46b5-91ac-81aaa244196a)
If array lengths are the same (I'd say at least 99% cases), I change
nothing in the code and the behaviour is exactly like before. When
lengths are different, aggregations didn't work, now they more or less
seem to do, so this PR should be only for good, and break nothing.

additional info:
Our response is still different than Elastic (look below, `-` instead of
our `0%`). They return `doc_count: 0` which is hard for us to do, as we
don't get such rows from Clickhouse. It's not completely trivial to fix
that, so I leave it for another PR.
![Screenshot 2024-06-17 at 16 30
29](https://github.com/QuesmaOrg/quesma/assets/5407146/260cea7c-a72b-4f7c-b947-6b9f85bbea39)