It should finally fix `mergeArrays` errors (#385)
So now basically whenever we split into buckets, there's this
`KeyAddedByQuesma` in each of those buckets.
Before it was added in a not always correct place.
It's still of course a bit of a hack, we only need this artificial key,
because we lose some information before during the processing. It can be
improved later. Now it at least should work in 100% of cases (minus this
`TODO` I left in the code, it should be an unlikely case though and I'll
fix that ASAP when I feel better, now coding is very painful for me)