Store query bodies non-concatenated
Before this change QueryBodyTranslated/translatedQueryBody were stored
as a single concatenated blob of all queries.

Change it to store each query body separately - [][]byte instead of
[]byte.

This change will allow subsequent improvements to the Quesma UI.

This change keeps the same behavior of the code (when
QueryBodyTranslated/translatedQueryBody gets actually used, it's
concatenated back to []byte).