Rewrite _resolve endpoint to use schema registry
Before this change, the _resolve endpoint had several bugs. One of them,
which affected me, was wrong result for an index present in Elastic,
but missing from ClickHouse and not enabled in the config. In that case
Quesma should reach to Elastic, but previously it just looked at
ClickHouse and immediately returned 404.

Since the previous logic was quite convoluted and used older LogManager
instead of schema.Registry, this commit largely rewrites the
implementation to be simpler, better documented and to use the
schema.Registry.

I extensively tested all possible scenarios (see the table in
Test_combineSourcesFromElasticWithRegistry), both as a unit test and
manually in Kibana.

The resolve code is now also moved to functionality/resolve package.