Rewrite _resolve endpoint to use the schema registry (#520)
Before this change, the `_resolve` endpoint had several bugs. One of
them, which affected me, was wrong result for an index present in
Elastic, but missing from ClickHouse and not enabled in the config. In
that case Quesma should reach to Elastic, but previously it just looked
at ClickHouse and immediately returned 404.

Since the previous logic was quite convoluted and used older
`LogManager` instead of `schema.Registry`, this PR largely rewrites the
implementation to be simpler, better documented and to use the
`schema.Registry`.

I extensively tested all possible scenarios (see the table in
`Test_combineSourcesFromElasticWithRegistry`), both as a unit test and
manually in Kibana.

The resolve code is now also moved to `functionality/resolve` package.

To make it possible to use the `schema.Registry` in the new code,
`schema.Schema` had to be extended with new boolean flag
`ExistsInDataSource`. Before the addition of this boolean flag, schema
registry contained both: schemas without created table in ClickHouse
(only specified/enabled in configuration) and schemas with created table
in ClickHouse - no easy way to differentiate between them.