Add initial support for "PUT /:index" integrated with schema.Registry
Add an initial support for the "PUT /:index" endpoint. This endpoint
can be used to set a static mapping for an index and this is what
the implementation focuses on.

After this change, a user can issue a "PUT /:index" before ingesting
the data and the created table will take that information into the
initial CREATE TABLE statement.

I have tested the solution with Flights and Ecommerce data and
the schema information inferred from "PUT /:index" is sufficient
to correctly ingest and query the data (removing all static
configuration!).