Execution plan (#554)
We are introducing `ExecutionPlan`. The idea is to keep everything
related to processing a single elastic query in a single place.

We currently keep only a list of queries to execute. 

There is a `QueryResultAdapter` added also. It will be used to implement
an alternative way of processing some aggregations. That adapter will
adapt alternative results set to the canonical one.