Rename index_statistics to ingest_statistics
index_statistics/IndexStatistics is only used for collecting statistics
about the ingest and it is NOT used for the query path. Rename the file,
struct to better describe what types of statistics it actually collects.

Doing this change now to prevent future merge conflicts and in some
preparation to potential future package reorganizations.