Fix spurious use of indexResolver preventing idling
We have noticed that Quesma still has problems with properly idling
(not making queries to ClickHouse Cloud to allow it to idle).

The current idling logic relies on checking if table discovery was
recently used. It turns out that idle Kibana very often does
inserts to an internal .kibana index and to /_monitoring/bulk endpoint.

Even though those requests are routed to Elastic, we still checked
if .kibana or _monitoring are names of enabled indices using
indexResolver - this use bumped the LastAccessTime, preventing idling.

Fix the issue by avoiding a use of indexResolver if it's an endpoint
(_endpoint) or if it's an internal index (starting with '.').