Support sending bulk entries to both Elastic and ClickHouse (#575)
_bulk requests can write data to many indexes. In particular, a bulk
request could contain writes to both enabled and disabled indexes, so
some parts of it should be sent to Elastic and some parts to ClickHouse.

Previously the code supported only either sending the entire request to
Elastic or sending the inserts to ClickHouse. After this change, the
code splits the bulk request into a Elastic and ClickHouse portions and
can send it to both endpoints.

The response from Elastic is parsed and "multiplexed" into the final
Quesma response (alongside results from ClickHouse). This required some
code refactoring and now the code has a proper support for setting
response codes/status per each inserted entry.