Pancake - update aggregation tests (#597)
Before we had only a few tests for pancakes, they were also in a
separate file `clients/ophelia_pancake.go`. Removed that file and added
a pancake version for every aggregation test.

Also a few-line fix for terms with empty buckets. Before empty buckets
case made rendering not work. A lot of tests failed because of that, so
it was much easier for me to fix it here.

---------

Co-authored-by: Jacek Migdal <jacek@quesma.com>
Co-authored-by: Rafał Strzaliński <rafal@quesma.com>