Replace Kibana Report with Kibana Discover in smoke test
Before this change, one of the smoke tests was a test of generating
a Kibana Report - a report of documents in an index. This tested
whether Kibana can really read from an index (via Quesma).

Unfortunately, Kibana internally generates this report by querying
the /_search endpoint ("*" wildcard endpoint). Previously it worked
by accident, because Quesma for wildcard "*" index happened to return
results from the first index, which happened to be the logs index.

Since now /_search is disabled, this commit introduces an alternative
way of making sure Kibana works correctly - it now uses an internal
search endpoint which is used when one uses the Kibana Discover UI
(code written based on observing requests it did).