Fix rendering of fields of array type in hits
Before this change, the fields of array type were not rendered correctly
in hits. For example, in the "Discover" tab in "Kibana Sample Data
eCommerce", the JSON representation of "manufacturer" would look like:

"manufacturer": [
  [
    "Elitelligence",
    "Oceanavigations"
  ]
],

instead of:

"manufacturer": [
  "Elitelligence",
  "Oceanavigations"
],

This could cause some arrays to display incorrectly in the Kibana UI.

The cause of the problem was that we always wrapped column values
into 1-element array, even if the value was an array itself. Fix
the issue by handling array values differently.