Add missing response body for "PUT /:index"
Before this change, the "PUT /:index" endpoint returned an empty
response (the response is not even mentioned in the documentation).

However, ElasticSearch Java Client (v7.17.23, used by the ElasticSearch
Sink Connector) parses the response and crashed upon seeing an empty
body.

Fix the problem by returning a proper response for this endpoint.