Pancake - `percentiles` metric aggregation (#630)
This one was very simple. I think that the only a bit tricky part is
```
if !strings.HasPrefix(res.ColName, "metric") && !strings.HasPrefix(res.ColName, "quantile") {
	// TODO remove second part of if after removing old aggregation logic!
	// In the new one, all quantile columns start with just "metric"
	continue
}
```
I wouldn't be sure if we don't detect some other metrics by that, so I
added a test for it, and we seem not to.