Fix validation of other numeric types (#636)
Before this change, if a user manually created a ClickHouse table with a
`Int32` column, the validation would reject inserts to that column as it
checked that `Int32` column != `Int64` value. Therefore the logic is now
extended to support all numeric types (larger/smaller/
floating-point/unsigned).

Additionally I added range checks that the inserted value actually fits
in the column (for example 1000 does not fit in UInt8 column).