Pancake - `(random_)sampler` aggr (#633)
If `(random_)sampler` is a root of the aggregation tree, we do a proper
sampling (add `LIMIT n`).
If it's somewhere in the middle of the aggregation tree, we remember
about it, so the result JSON is correct, but we simply omit the
sampling, DB queries are just like they would be without the sampler.

Also `random_sampler` behaves now in the same way as `sampler`.
It can probably be easily improved using
https://clickhouse.com/docs/en/sql-reference/statements/select/sample.
But let's do that in another PR

---------

Co-authored-by: Jacek Migdal <jacek@quesma.com>