Add "get mapping" endpoints (GET /:index, /:index/_mapping)
This commit adds support for GET /:index and GET /:index/_mapping
endpoints. The primary use of those endpoints is to get the (static)
mapping of an index. Those endpoints are actively used by some Elastic
clients, such as ElasticSearch Kafka Sink Connector.

The implementation generates the mapping based on the schema found
in the schema registry.

hierarchical_schema.go was introduced since the schema registry
stores the schema in a "flat" format (e.g. "products.product_id",
"products.base_price"), but the mapping we need to return should be
nested. The functions in that file transform the flat schema into a
tree, which is more convenient to use.

The added tests in mappings_test.go show that a round-trip is possible:
sending a mapping to Quesma, it getting parsed and stored as a Schema
struct and generating back a mapping based on that representation.
(It's not 100% equal, since we change numeric types to wider types, for
example floats to doubles.)