treewide: Pass QuesmaConfiguration by pointer, not by value
While working on #683, I noticed that the code in `process()` in
Statistics modified QuesmaConfiguration struct. However since
QuesmaConfiguration was passed by value, this modification
essentially did nothing.

That (attempted) modification of global state is not that clean
(fixed in the next commit), but this highlighted an issue: the
entire codebase always passes QuesmaConfiguration by value (and
this is quite a big struct), not by pointer. This also makes
it not possible to change the configuration in the runtime
(since everyone is using their own copy).

Fix the issue by modifying all uses of QuesmaConfiguration to pass
it as a pointer.