Send proper error response in `/_doc` endpoint (#708)
Before this change, the `_doc` endpoint would always report that the
insert has succeeded, even if an error occurred and couldn't be handled
by Quesma (meaning the insert failed!).

The `RunConfigured` function previously knew about the error, logged it,
but did not pass it up to the caller. Change that and return a properly
formatted error response in `doc.go` (`bulk.go` already handled it
correctly).