Move IgnoredFields into SchemaOverrides, make it work properly with ingest (#701)
This is a part of our recent effort of finalizing a new stable
configuration. Previously `IgnoredFields` was a part of
`IndexConfiguration`, looking like this:

```yaml
test_index:
  ignoredFields:
    myField: true
    anotherField: true
```

However, with the recent improvements to `SchemaConfiguration`, now it
makes more sense to configure it under `schemaOverrides`, like this:

```yaml
test_index:
  schemaOverrides:
    fields:
      myField:
        ignored: true
      anotherField:
        ignored: true
```

This PR moves this configuration from `ignoredFields`
(`IndexConfiguration`) to `schemaOverrides` (`SchemaConfiguration`).

Moreover, `ignoredFields` did not work correctly with Quesma's ingest,
as ignored fields were only ignored in the query path. This PR adds
necessary checks also to the ingest path.

Unfortunately, the code is not very clean with regards to handling of
`.` vs `::` - I have tried cleaning up the situation, but that required
some larger refactors which would make this PR too big...