Push Elastic's response from A/B testing to management console
In the Quesma Managment Console we have "Elastic response" panel
in the "Live tail" page and similar on the drilldown page. Those panels
were filled in by the "EnableElasticsearchIngest: true" mode, which
resends all requests to Elastic (even non-ingest!).

However in the upcoming PRs we will effectively remove
"EnableElasticsearchIngest" mode. For the query path, it's now obsoleted
by A/B testing with Elastic (see #631). But A/B testing with Elastic
did not fill in the Quesma Management Console with responses from
Elastic.

This PR improves A/B testing with Elastic to also send Elastic's
response to the Quesma Management Console.