Remove EnableElasticsearchIngest
Remove EnableElasticsearchIngest option. This mode essentially always
resent each request to ElasticSearch. Contrary to its name, it resent
it for both query path and ingest path. This re-sending happened on
a low-level - routing logic. It turns out this is too low-level for
query and ingest needs:

Query needs:
A/B testing with Elastic (a newer approach) is already implemented
and achieves more than could be done at the low-level routing level
(comparisons, only doing it for search endpoints, not other endpoints)

Ingest needs:
EnableElasticsearchIngest primary use-case was double writes to Elastic
and ClickHouse, but this flag was global, not configured per each index.
In the upcoming PR we will reimplement double writes at a higher level,
not at the routing level ("dumb" replaying of HTTP requests)