Introducing internal Quesma type (#732)
We've found the `Type` name being very elusive in our codebase when
working on schema persistence with @nablaone.

This PR introduces `QuesmaType*` so that it becomes very clear in our
codebase when we're referring to ClickHouse/Elastic types vs our
internal ones.

Additionally, `config.QuesmaConfiguration` is no longer used in
SchemaRegistry in favor of just index configuration.